project/backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.2.1
